requirements.txt
python-dotenv
streamlit
langchain-core
langchain-groq
PyMuPDF
langid
pandas
scikit-learn==1.2.2
numpy==1.26.4
deep_translator